requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0

# LangChain and LangGraph
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-google-genai>=3.1.0
langgraph>=1.0.0
langchain-core>=1.0.0
google-generativeai>=0.8.0

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Utilities
schedule==1.2.0
python-dateutil==2.8.2

# Image processing
Pillow>=9.0.0

# Logging and monitoring
loguru==0.7.2

# Data handling
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
